move save preferences to controller
